requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0
scipy>=1.7.0
albumentations>=1.0.0
wandb>=0.12.0
scikit-learn>=1.0.0
seaborn>=0.11.0
openpyxl>=3.0.0
